@@ -22,45 +22,45 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^6.0.3",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/pwa": "^0.6.8",
-    "@angular/router": "^6.0.3",
-    "core-js": "^2.5.4",
-    "rxjs": "^6.0.0",
-    "zone.js": "^0.8.26",
-    "@angular/service-worker": "^6.0.3"
+    "@angular/animations": "^9.0.2",
+    "@angular/common": "^9.0.2",
+    "@angular/compiler": "^9.0.2",
+    "@angular/core": "^9.0.2",
+    "@angular/forms": "^9.0.2",
+    "@angular/http": "^7.2.16",
+    "@angular/platform-browser": "^9.0.2",
+    "@angular/platform-browser-dynamic": "^9.0.2",
+    "@angular/pwa": "^0.900.3",
+    "@angular/router": "^9.0.2",
+    "core-js": "^3.6.4",
+    "rxjs": "^6.5.4",
+    "zone.js": "^0.10.2",
+    "@angular/service-worker": "^9.0.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular-devkit/build-ng-packagr": "~0.6.8",
-    "@angular/cli": "~6.0.8",
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular/language-service": "^6.0.3",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "~0.900.3",
+    "@angular-devkit/build-ng-packagr": "~0.900.3",
+    "@angular/cli": "~9.0.3",
+    "@angular/compiler-cli": "^9.0.2",
+    "@angular/language-service": "^9.0.2",
+    "@types/jasmine": "~3.5.5",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~13.7.4",
+    "codelyzer": "~5.2.1",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "lite-server": "^2.4.0",
-    "ng-packagr": "^3.0.0-rc.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tsickle": ">=0.25.5",
-    "tslib": "^1.7.1",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
+    "karma-jasmine": "~3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.2",
+    "lite-server": "^2.5.4",
+    "ng-packagr": "^9.0.1",
+    "protractor": "~5.4.3",
+    "ts-node": "~8.6.2",
+    "tsickle": "0.38.0",
+    "tslib": "^1.10.0",
+    "tslint": "~6.0.0",
+    "typescript": "~3.7.5"
   }
 }
